Bunch of cleanups in common code

+++ hmf-common/src/main/java/com/hartwig/hmftools/common/genome/refgenome/MockRefGenome.java
@@ -3,7 +3,6 @@ package com.hartwig.hmftools.common.genome.refgenome;
 import java.util.Map;
 
 import com.google.common.collect.Maps;
-import com.hartwig.hmftools.common.genome.refgenome.RefGenomeInterface;
 
 // test implementation of the ref genome
 public class MockRefGenome implements RefGenomeInterface
@@ -21,9 +20,8 @@ public class MockRefGenome implements RefGenomeInterface
         String chrBases = RefGenomeMap.get(chromosome);
 
         if(chrBases != null && chrBases.length() > posEnd)
-            return chrBases.substring((int)posStart, (int)posEnd + 1);
+            return chrBases.substring(posStart, posEnd + 1);
         else
             return "";
     }
-
 }
